+++ java-design-processor/src/test/resources/internal/nbbrd/design/ValidSFM.java
@@ -3,13 +3,27 @@ package internal.nbbrd.design;
 import nbbrd.design.Immutable;
 import nbbrd.design.StaticFactoryMethod;
 
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+
 public class ValidSFM {
 
-    @StaticFactoryMethod
-    public static ValidSFM make() {
-        return new ValidSFM();
+    public static class C1 {
+
+        @StaticFactoryMethod
+        public static C1 make() {
+            return new C1();
+        }
+
+        private C1() {
+        }
     }
 
-    private ValidSFM() {
+    public static class StreamUtils  {
+
+        @StaticFactoryMethod(InputStream.class)
+        public static InputStream nullInputStream() {
+            return new ByteArrayInputStream(new byte[0]);
+        }
     }
 }